Merge kevin_dev into master (#27)
* Can't force RPATH with static link.

* Fixing the CMakeLists.txt file.
The CMakeLists.txt file does too many non-standard things that cause
portability problems (especially on ORNL resources). In order to make
sure that builds still work in spack and on other systems, we need
to revamp the CMakeLists.txt file.  this commit is the first step
on that process.

* Cleaning up search for EVPath.

* Cleaning up build.
Fixing MPI dependencies, EVPath dependencies, Python dependencies.
Automatically build the Python library if requested.

* Fixing target dependencies for python build
And making python build quiet.

* Don't enable EVPath cloud by default

* Don't make ssos_python.c a source file.
It is generated.

* Added Munge cmake support, fixing config bugs.

* Making everything consistent.

* Clarifying options in the top of main CMake file.

* Fixing sanitizer options.

* Minor changes for OSX support.

* Adding missing gcc pragma push

* Merged SOS_feedback_type into SOS_msg_type.